Digibooks Code Repository
This contains the code repository for digibooks.com.
For this you need to have Apache Tomcat version 8.0 installed and
running on your system. After that you also need to provide reference to
Oracle JDBC driver in the project build path and Deployment Assembly.
Very Very Important : Do not commit any changes unless you are
absolutely sure that the changes you have done are correct and working
since this is a common code repository